Split Homepage and SecondPage into separate components
Also renamed SecondPage to FitnessTypesPage to be more descriptive.

Right now the web app functions the same as it did before, but now different pages are split into different components. From my experience, the sooner you split different parts of your web app into discrete components, the less time you will spend later separating them when they become more complex.

In App.jsx:
We return a Router component that points to the path "/" by default. Therefore, our HomePage is rendered by default because the web app URL is something like localhost:5173/ by default.

In HomePage.jsx:
When we click on the "My Metrics" button, the function navigate('/fitnessTypes') is called, changing the URL to localhost:5173/fitnessTypes. The router in App.jsx sees this path change and then returns the FitnessTypesPage component.

In FitnessTypesPage.jsx:
We see our list of fitness types as well as a "Back to Home" button that changes the URL to localhost:5173/ so that our homepage is reappears.