Metric Display for Weight Lifting - More Graphs Added
Added graphs for:
- Total Reps
- Total Volume
- 1 Rep Max
- Strength to Weight Ratio
- Avg and Max Heart Rates
- Estimated Calories Burned

There were a fair few calculations used to estimate the calories burned for a given weight lifting session. What's nice is that we should be able to reuse these calculations for any kind of physical activity given the proper information.

If you are curious, my comments hopefully explain the math.