Renamed folder: components to pages
This is because these are pages, not individual components like a group of text fields or a graph. Each page may eventually contain such components, but I'll bet we will eventually add discrete components that can be loaded on any page. They will be stored in a components folder when we add some