Inputting Running Metrics now works
Added the ability for the user to open a menu full of text fields where they can input all relevant metrics for a given run. The graphs above will update as the user inputs more data into them.

The implementation for this maybe could be slimmed down eventually, but here's a basic explanation for how it works. I re-declared the array of objects called runningData as state, so that changes in its value would appear on screen without refreshing the page. Each new text field is linked to its corresponding state so that their values can update in the same way. Finally, clicking the Submit button pushes a new object with the appropriate state into runningData. From there the graphs populate with their data.

There is currently a mathematical bug where you can burn negative calories in your exercises. As a user, I would be cross if I went for a run and my body thought I ate a piece of chocolate cake instead. Needless to say, I will be looking into this very soon!