Added Swimming Metric Input
This was by far the most complicated metric to provide input capabilities for. Since lap times and stroke counts per lap depend on the number of laps a swimmer will complete. I made it create N text fields for each lap times and stroke counts based on the number of laps. If that explanation isn't worded the best, I would pull it down and try it for yourself :)