Basic Set Up of Hono connecting to MongoDB
This is an experimental branch currently. I recommend downloading it and following the video tutorial I linked.

If you open your terminal in visual studio code and type the command: cd api
then type the command: npm start
You will start a local server connection using Hono Server. Much like our client app, you can open it in your browser by visiting: localhost:3000

Once you open the API in your browser, you will see the text, "Hello Hono!". This is because of this code in our app.js within the api directory:
// Our base home route
app.get('/', (c) => {
    return c.text('Hello Hono!')
})

Also included is a model (similar to a table in SQL) called Running. It currently defines the data that will be stored for a running exercise. We are still going to have to come to a consensus on the design of our database, so this model is more for testing. In fact this entire branch is open for testing until we start using our API (Hono) to add to our database.

Included are all of the dependencies. If you have issues, try the command: npm install