You can now input data directly to the database
This is all for testing purposes, and much of this is subject to change. I have defined a collection in api\models\RunningExercise.js that gets imported into api\app.js to be used for posting content to that same collection. Note: For some odd reason, the collection in MongoDB is listed under the "test" database and is called "runnings" when I defined it as "Running". This is because I didn't follow the REST API URI naming conventions. I tried to name a collection after a verb, which is not permitted. Here is a link to those conventions (they are more important than I realized): https://restfulapi.net/resource-naming/

In client\src\pages\RunningMetrics.jsx, I have replaced the contents of this file with the version from the main branch that allows for user input. I have also replaced the body of the handleSubmit function. It still checks for erroneous data, but now creates an object which gets posted to the path "http://localhost:3000/exercises/running".

If you then go to api\app.js you can see where I have defined the behavior for a post call at that same path (here it's shortened to "exercises/running". The Input is basically converted to JSON and then fed into the create function to create a new record in MongoDB with the data found within our JSON object (body).

I realize the name of this branch may be misleading. If I could, I would rename it to Database-Experimentation because a lot of this will change as we progress.

I also ran into an issue where I had to downgrade react and react-dom to version 18.3.0. This is because version 19 isn't a stable release. To do this, all you have to do is change the version names in package.json, then delete package-lock.json and the node_modules folder. After that, just run the command: npm install.