Graphs now pull data from MongoDB!
I added a useEffect hook to  RunningMetrics.jsx that fetches data from the "runningentries" collection in the "test" database. Once a new exercise is submitted, a variable called "fetchCount" is incremented, causing the hook to run again, fetching the updated data for the graphs to display.

I also fixed a small oversight with the isError function that prevented the user from entering in fractions of a mile or hour.

Also renamed the collection "Running" to "RunningEntries"