Small Changes to our home page and fitness types page
I removed the proof-of-concept light/dark mode toggle button and all associated styling. It all lives in theme.jsx for future referencing. In that file I started to define a theme, but thought it would be easier to temporarily work within main.jsx instead.

We are now importing Links as RouterLink components and MuiLinks from MUI. Basically, we can only override the styles of MUI components, so I found a way to import MUI's version of a link, but keep the functionality of react-router-dom's Links.

It was hard not to have my styles applied to other pages. I know some of us have tasks to apply the standard theme and other styles to the pages that we worked on, but due to the nature of theme overriding, it all gets applied automatically. I've left as much of the styling alone as possible to give other's a chance to do some themselves.