More RunningMetrics Cleanup and Formatting
I chose to standardize these three components so that they are easy to compare. One of the main things is making sure all statements are in the same order, and no excess stuff is included. Hopefully this helps us in the future if we ever have to go in and troubleshoot some elusive bug!