Added an exercise schema and API behavior
This is the exercise schema that I showed during our meeting this afternoon. Based off the design for our database, all exercises added to the database are being added to the same table or collection. The "type" of each exercise is used as a parameter for querying recorded exercises. In other words, if we want to query all running exercises, we provide our async getRunningExercises function in ViewRunningMetrics.jsx with the parameter: type: "run". This filters out all exercises that aren't running exercises. Eventually we will be able to add more parameters such as userID.

In short, we are now recording all our exercises in localhost:3000/exercises instead of localhost:3000/exercises/running-entries.