All other metrics/exercises are connected to the DB
Now each metric page's data is persistent and specific to the type of exercise entered. In other words, only swimming exercise metrics are going to be displayed on ViewSwimmingMetrics.jsx.

like I had mentioned last commit, we are currently not filtering by user. Another thing we are missing is our implementation of dates. We are unable to see within the app WHEN an exercise was recorded, but all that data fortunately lives in our DB since it is automatically generated!