Added a Record Exercises page
This is a dedicated page for entering any exercise of any type into the database. I've tested it and it seems to work as expected. You first select the exercise type and then the corresponding Input exercise component gets loaded for you to actually record an exercise into the database!

There is still some weird styling inconsistencies that I would like to get to the bottom of. I would also like to alter the existing metrics pages so that they are only for viewing metrics.