Added some rudimentary Global State!
There is still a lot I would like to clean up here, but you can now log in as either Bob or Joe and their global state is accessible in: HomePage (console.log() test outputs), ViewRunningMetrics, and InputRunningExercise. For the last two in this list, you are able to view the running metrics for Bob or Joe and even record a running exercise as Bob or Joe.

When logging in, all user data is pulled down and a new handleSubmitTest function looks through the array of all users (just Bob and Joe right now) and uses the find method to look for a username and password match. Once found, the functions selectUser and setTheme are called. These refer to the reducer function defined in GlobalStateContext.jsx where given an action.type, an action.payload (the data) is assigned to a given part of the initialState.

This stuff is crazy, but it allows us to do so much! I'm going to clean this up and document more.