Fixed a bug that broke the dark mode theme
When your computer is set to light mode (or dark mode for that matter), the CSS color-scheme property is used to inform the browser about the color themes your browser supports.

So I suppose this is why we had a white background with the dark theme selected while my computer was set to light mode. I basically had to disable the color-scheme property in main.jsx on the same level in the DOM as our entire App (where it is being wrapped by our ThemeProvider).

So no longer will components like text detect their background and automatically stand out in contrast. We are going full manual mode here with the styling, which I hope I don't regret. At least we are given full control here.