Now your system preferences determine the theme
This is documented quite well so I would read the comments if you are curious.

This is leading me to consider if we really should be storing the user's theme preference in our DB. A toggle is fun and all, but I (as a dark mode enjoyer) never want to see a webpage or application blind me with a light mode theme. Thus why I set my system preferences to dark mode to begin with.