Added dynamic routing and global state
This page is now dynamically routed, meaning in the url instead of '.../profile' its '.../profile/id' so you can choose to look at any page.
Also connected it to global state so you will only see the option to edit if you are logged into the profile you are viewing. 

One issue remaining is that if you go to a profile page that doesn't exist the page is blank. We should add some error handling to fix this. 

Also fixed some styling things that had been fixed previously, but must've been undone during a pull request at some point.