Applied Theme Styling to the Profile Page
I just changed an h3 component to a Typography component.

Other than that I was trying to fix a spacing issue with the Friends Card. It touches the edit Card and was trying to apply the same spacing to it. It looks like I did a lot here, but I was trying to get a better understanding of the component structure (what wraps what) by applying some better tabbing to things that are nested to improve readability.