Minor Fixes to Diet Plan
Replaced the Paper component that contains all diet info with a Card. There was a white background with white text on top with our dark theme. Another small change is now the Card won't render with all the contents of the diet plan UNTIL you select a diet. The user will be instructed to select a diet first.