Made Graphs and Images smaller
This was a compromise I settled on because there were two many variables at play for me to apply the responsive design I wanted. I was comparing the app on my desktop monitor and my laptop display, and everything is still significantly larger on my laptop. I was able to get some nice responsive design to work by wrapping the graphs in a Box component, however this completely broke the graphs. They don't seem to like being dynamically resized as a percentage of their wrappers' dimensions.

I settled on slightly changing the dimensions of the graphs:
from 500x300 to 400x250