Added the Toolbar back to the HomePage
This is more of a placeholder than a finished toolbar. I decided to keep it for now so there's not too much confusion when merging branches this sprint. However, it shouldn't have been merged in the first place because it A: wasn't approved by anyone else and B: doesn't fulfill the acceptance criteria of the associated Jira task.