Achievement Schema and API querying
Added a new Achievement mongoose schema so I could define query behavior for our API. The plan is to pull all achievements from the database and compare their criteria to all the user's exercise data.  So it turns out that we will also need to have the exercises array for each user filled out. We could also make another query to our exercises table, filtering by userID, but I wanted to see if I could minimize our API calls as much as possible. This is far from an optimized system because we aren't caching any query calls. However, I might look into that and earn us a few extra brownie points on this project if I have time this week. That could also be a task for next sprint.