Moving code from my laptop
This commit has some unfinished code because I'm just using it to move code from my laptop to my desktop.

I realized that it would be really nice to have all exerciseID's stored in the exercises array that each user has, so I am working on updating that as new exercises are recorded. That way I can easily make references to past exercises by ID when checking for achievement completion on the profile page.