Achievements now display on a user's profile page!
Now when you visit a user's profile, you can see which achievements they have earned. They will be displayed in a Card component as individual Chips, all under the bio. The styling/spacing could use some work, but it is functional.

Note: This also works if you switch to someone else's profile page.

There are a few things I've learned about asynchronous fetching and async functions as a whole, and I've taken the time to add some comments to cement some of this new knowledge. If you are interested, it may be helpful to read some of them!