Removed the Achievements button from HomePage
The old achievement page walked so the profile/achievement could run! We no longer need the old page, so I removed the link from the home page. But I kept RewardAchievements.jsx so we can refer back to it.

We also had a rewardachievements.jsx on the same level as our App.jsx??? Got rid of it because it was empty