User taken error message + ability to edit weight
Added a username taken helper text that will appear when attempting to create an account for a user that already exists. If no username or password is entered then the box turns red but has no helper text.
Added the ability to edit weight from the profile page so instead of providing basic information when you create your account you can just go to your profile and edit stuff there.