Updating weight and checking for existing account
For profile page fixed the way a new weight was added to the weightArray (it was becoming a nested array of arrays, rather than just the array of objects).

For create account fixed the way it checked for an existing account. Previously had a bug that would display the text that that account already existed but would still create the account.