Swapped back to using the Exercises table
Before, I was using the exercise array belonging to each user to query exercises with the ID's it contained. This relied on additional global state that is more convoluted than it needs to be. This is still a work in progress, but I will be working to clean up some of the things we no longer need in subsequent commits.