Added Tooltips to each achievement chip
This way you can hover over an achievement and see what it took to earn it! In addition to adding a tooltip to every achievement in our MongoDB achievement table, I added and changed all of the achievements. I changed names and added new criteria. Go ahead and take a look if you're curious :)