A few adjustments elevationGain achievements
Since elevationGain is a metric shared by Hiking and Cycling, I wanted to make sure we were distinguishing between both activities. For example, on my profile, I have hiked over 2000 feet in elevation, but I previously got the achievement for cycling 2000 feet in elevation. This is because I was previously not separating this stat by activity.