All user exercise data has been fed in and formatted
I'm going for a clean slate with this version of the leaderboard page, so I have deleted everything and starting fresh because this is one of our oldest pages.

All the new code you see here formats user metrics very similarly to how they are formatted on the profile page for determining which achievements a user has earned. The format is a JSON object that will look something like this:
object = {
    distance: {run: 90.3, cycle: 5, hike: 10}
    duration: {run: 16.150000000000002, cycle: 0.75, hike: 2}
    elevationGain: {cycle: 250, hike: 500}
    elevationLoss: 0
    lapCount: 8
    maxHeartRate: 141
    maxWeightOfWeights: 8
    name: "Crow"
    steps: 121200
    totalLapTime: 355
    totalReps: 270
    totalStrokes: 441
    totalVolume: 1710
}

I have an array of these objects, so hopefully I can form a meaningful leaderboard with this data.