The leaderboard now sorts and categorizes metrics
First, we populate an object called "leaderboard". It has some commented out parts that show the kinds of objects that will be stored in its arrays. Then we sort that leaderboard object so that we can easily determine who is in the lead for a specific metric.

Next I will find a way (probably graphs) to display these metrics and the leaderboard as a whole in the app!