Tentative Button Layout
There is one button group that allows the user to select which category of exercise they want to select. Depending on which category they select a second button group will appear and populate with corresponding metrics.

Ex: You click the Running button and three more buttons appear "Distance", "Duration", and "Steps".
Clicking any of the buttons in the second group will result in the appropriate metrics being outputted to the console.

I'm calling this tentative because I might make a buttonGroup component that I can feed data to. Currently there is a lot of repeated code/components. But for now I'm going to focus on functionality before making my code easy to read.

Next is to add a visual component to these metrics!