Fixed a bug where the graph wouldn't load
If either of the three are undefined, we start to run into problems:
- selectedMetric
- exerciseCategory
- exerciseMetric

Also fixed a smaller bug where if you swapped between categories, the graph would persist with a metric from the previous category. I fixed this by adding to the dependency array of the useEffect hook on line 291.