Added additional text fields
Now when you create an account, you enter your username and password, and if they are both valid and available, a new series of text fields appear. Text fields for name, height (ft. and in.), weight, and description appear and a separate function will handle the error-checking of those fields. I will most likely borrow from the editing profile info code when it comes to such checking. Once It is determined that the user has entered valid input for each step of the account creation process, then the createUser function will be called and the accountCreated state will be set to true, prompting the user to log back in with their newly-created account. That's the plan at least!

Other changes:
- Now working with the test database until this feature works 100% as intended
- Got rid of some excess console.log statements in LoginPage.jsx