All user Info is now error checked
Borrowing from the existing conditions in the profile page's edit logic, I was able to implement error checking for user info as well as some helpful error messages! I'm going to look into a few things relating to profile pictures now. I would like the user to be able to choose their profile picture from this menu as well as possibly be able to preview it.

Other Changes:
- Changed the minimum weight from 0 to 1 in ProfilePage.jsx