Accessing empty profile redirects
Now when you try to access /HomePage/profile/ it shows the log in page. This fixes the issue that would occur when you try to access 'my profile' from the all profiles page or from the future tool bar. 
(Accidentally committed this file to the wrong directory before which is why I'm recommitting it now)