Formatted all exercise data by date into JSON
Here is an example of an array of JSON objects for Alex:

0: {date: '2025-03-25', count: 11, level: 1}
1: {date: '2025-04-08', count: 1, level: 1}
2: {date: '2025-04-10', count: 1, level: 1}
3: {date: '2025-04-14', count: 1, level: 1}
4: {date: '2025-04-17', count: 1, level: 1}

This is the format that works well with the react activity calendar. I will just have to toy around with the level key later on. Next commit I'm going to try and add this library and see how it looks!