Fixed a bug with date ranges
The comment really says it all. Basically the array of objects containing ISO-8601 dates needs to be sorted in ascending order by date because the react activity calendar library looks at the first and last dates in the array to determine the range of the calendar. You could imagine that dates could fall out of this range and never be displayed.

It turns out that the date-fns library has some sleek functions for parsing these dates in a standard ISO-8601 format so that compareAsc can sort them in ascending order!

Also removed some extra console logs.