Manually adjusted the calendar range
The React Activity Calendar will use the first and last dates of the data given to it to determine the range of the calendar. So before I add any more data, I'm adding a dummy date for 1-1-25 so that the whole year to date shows on the activity calendar. (Note that this will start to look funny when 2026 rolls around). I am also pushing dummy date for the current day if the user didn't already exercise today.

Today is 5/1/25, so as of today, the activity graph will show dates from 1/1/25 to 5/1/25. That was my intent here :)