Added a date range slider to ViewHikingMetrics
This new slider allows you to view all hiking exercises in the graphs that fall within a specific date range. The purpose of this was to eliminate issues when viewing metrics pages with a lot of exercises recorded (more than 15). The graphs are of fixed size, so the individual chart bars only get smaller with more data. This slider allows the user to "zoom in" to see their data more easily.

Next I will be adding this to all other View Components!