Fixed another bug preventing users from updating their profiles
This one really came out of the blue and I'm glad I caught it! Now since we are auto-filling the input fields with the result of these map functions, we are essentially setting the text fields to a value contained in an array. This is because the map function returns an array. So when we would edit the profile and change just one or a few text fields, the remaining fields held the correct data, but inside an array. Our database schema wasn't built to deal with arrays, so the ability to edit profiles silently failed.

All I am doing is indexing to the first (and only) element of these arrays returned by map before assigning them to the state that is connected to the value of the text fields.